fix : asteroid muter & mencong
